Move searchItemsByString to wc-api/items/utils.js
